I uh, kinda rewrote most of the mod I think.
CelesteNet integration now only sends updates on connecting, when respawning, and when changing selected hat. And when a hat is forced by a trigger.

The CelesteNetHatComponent now iterates over Ghosts in the Scene and just gives them a default HatComponent; the component then detects that its Entity is a Ghost and gets the hat data out of a BoundRef